now done a transactional method to delete togheter with annunci, the spedizione

+++ src/main/java/com/example/TrasportiBackend/Annuncio/Annuncio.java
@@ -22,7 +22,7 @@ public class Annuncio {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
     private LocalDate dataPubblicazione = LocalDate.now();
-    private long retribuzione;
+    private int retribuzione;
     @ManyToOne
     @JoinColumn(name = "azienda_id")
     private Azienda azienda;
