+++ src/main/java/com/example/TrasportiBackend/payloads/entities/AnnuncioDTO.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull;
 
 public record AnnuncioDTO(
         @NotNull(message = "Retribuzione necessaria")
-        long retribuzione,
+        int retribuzione,
         @NotNull(message = "Azienda id necessario")
         long aziendaId,
         @NotNull(message = "Spedizione id necessario")
